refactor(web): unify footer across all non-paginated pages (ARC-692)
Replace the per-page footer wiring with a single layout-level footer:
- promote the home-page footer markup/styles into the shared
  `widgets/footer` widget (renamed `AppFooter`, CSS lives next to it)
- render it from the root layout via a new `LayoutFooter` client wrapper
  that hides it on pages that own their pagination (/games, /history,
  /notes, /leaderboards, /stats)
- drop the now-redundant `AppFooter`/`HomeFooter` imports from
  HomePage, SettingsContent and SupportPage
- retarget the existing footer e2e specs at the new shared widget
