+++ apps/web/e2e/footer-links.spec.ts
@@ -11,7 +11,7 @@ test.describe('Footer Links', () => {
     await navigateTo(page, '/');
     await page.waitForLoadState('load');
     // Ensure the footer is attached to the DOM
-    const footer = page.locator('footer.home-footer-root').first();
+    const footer = page.locator('footer').first();
     await footer.waitFor({ state: 'attached' });
 
     // Scroll to the bottom
@@ -23,7 +23,7 @@ test.describe('Footer Links', () => {
 
   test('should show copyright information', async ({ page }) => {
     await scrollToFooter(page);
-    await expect(page.locator('footer.home-footer-root').first()).toContainText(
+    await expect(page.locator('footer').first()).toContainText(
       /©|2026|Arcadeum/,
       {},
     );
@@ -31,7 +31,7 @@ test.describe('Footer Links', () => {
 
   test('should show app version', async ({ page }) => {
     await scrollToFooter(page);
-    await expect(page.locator('footer.home-footer-root').first()).toContainText(
+    await expect(page.locator('footer').first()).toContainText(
       /\d+\.\d+\.\d+/,
       {},
     );
@@ -42,7 +42,7 @@ test.describe('Footer Links', () => {
   }) => {
     await scrollToFooter(page);
 
-    const footer = page.locator('footer.home-footer-root').first();
+    const footer = page.locator('footer').first();
     // On mobile, footer sections are collapsible - expand Legal section if needed
     const privacyLink = footer.locator('a[href="/privacy"]');
     if (!(await privacyLink.isVisible())) {
@@ -72,7 +72,7 @@ test.describe('Footer Links', () => {
   test('should show social or support links', async ({ page }) => {
     await scrollToFooter(page);
 
-    const footer = page.locator('footer.home-footer-root').first();
+    const footer = page.locator('footer').first();
     // On mobile, footer sections may be collapsed - expand "Follow Us" if needed
     const socialLinks = footer.locator('[data-testid^="footer-social-"]');
     if (
