+++ apps/web/src/app/home/HomePage.tsx
@@ -3,8 +3,8 @@ import HomeHero from './components/HomeHero';
 import { DailyRewardChip } from '@/features/daily-rewards/ui/DailyRewardChip';
 import dynamic from 'next/dynamic';
 
-// Single concatenated bundle of hero + footer + presentation styles.
-// The 3 originals each became a separate render-blocking chunk under the
+// Single concatenated bundle of hero + presentation styles.
+// The originals each became a separate render-blocking chunk under the
 // Lighthouse simulator's per-chunk model (~303ms penalty each). Bundling
 // to a single physical file collapses them to one HTTP round-trip.
 // See docs/superpowers/specs/2026-05-06-home-perf-phase-2-diagnostic-results.md
@@ -18,7 +18,6 @@ const HomePitchDeck = dynamic(() => import('./components/HomePitchDeck'));
 const InstallAppCta = dynamic(() =>
   import('@/widgets/install-app').then((m) => m.InstallAppCta),
 );
-const HomeFooter = dynamic(() => import('./components/HomeFooter'));
 
 export default function HomePage() {
   return (
@@ -33,7 +32,6 @@ export default function HomePage() {
       <HomePresentation />
       <HomePitchDeck />
       <InstallAppCta />
-      <HomeFooter />
     </PageLayout>
   );
 }
