+++ apps/web/src/app/home/components/styles/home-bundle.css
@@ -325,338 +325,6 @@
     background: rgba(30, 32, 35, 0.92);
   }
 }
-/* Static Footer Styles */
-
-.home-footer-wrapper {
-  width: 100%;
-}
-
-.home-footer-root {
-  width: 100%;
-  background-color: var(--glass-background, rgba(0, 0, 0, 0.4));
-  border-top-width: 0;
-  padding-top: var(--t-space-12, 48px);
-  padding-bottom: var(--t-space-10, 40px);
-  position: relative;
-  backdrop-filter: blur(32px) saturate(180%);
-}
-
-.home-footer-border-line {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 1px;
-  pointer-events: none;
-  background: linear-gradient(
-    90deg,
-    transparent 0%,
-    var(--glass-border, rgba(255, 255, 255, 0.1)) 15%,
-    var(--primaryGradientStart, #81f1ff) 50%,
-    var(--glass-border, rgba(255, 255, 255, 0.1)) 85%,
-    transparent 100%
-  );
-  opacity: 0.8;
-}
-
-.home-footer-container {
-  width: 100%;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding-left: var(--t-space-4, 16px);
-  padding-right: var(--t-space-4, 16px);
-}
-
-.home-footer-grid {
-  display: flex;
-  gap: var(--t-space-12, 48px);
-  flex-wrap: wrap;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.home-footer-brand-col {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-6, 24px);
-  flex: 2;
-  min-width: 320px;
-  max-width: 600px;
-}
-
-.home-footer-brand-info {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-4, 16px);
-}
-
-.home-footer-brand-name {
-  font-size: 24px;
-  font-weight: 800;
-  background-image: linear-gradient(
-    135deg,
-    var(--primaryGradientStart, #81f1ff) 0%,
-    var(--primaryGradientEnd, #8198ff) 100%
-  );
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  margin: 0;
-}
-
-.home-footer-description {
-  font-size: 16px;
-  color: var(--text-muted, rgba(255, 255, 255, 0.6));
-  line-height: 1.5;
-  max-width: 500px;
-  margin: 0;
-}
-
-.home-footer-social {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-4, 16px);
-}
-
-.home-footer-social-title {
-  font-size: 12px;
-  font-weight: 700;
-  letter-spacing: 0.1em;
-  margin: 0;
-  color: var(--text-primary, #ffffff);
-}
-
-.home-footer-social-links {
-  display: flex;
-  gap: var(--t-space-3, 12px);
-  flex-wrap: wrap;
-}
-
-.home-footer-social-icon {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  border-radius: 16px;
-  background-color: rgba(255, 255, 255, 0.05);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  color: rgba(255, 255, 255, 0.5);
-  text-decoration: none;
-  transition: all 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275);
-  position: relative;
-  overflow: hidden;
-}
-
-.home-footer-social-icon:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-  border-color: var(--primaryGradientStart, #81f1ff);
-  color: #ffffff;
-  transform: translateY(-8px) rotate(15deg) scale(1.15);
-  box-shadow:
-    0 15px 30px rgba(0, 0, 0, 0.5),
-    0 0 30px rgba(87, 195, 255, 0.3);
-}
-
-.home-footer-social-icon::after {
-  content: '';
-  position: absolute;
-  inset: 0;
-  background: radial-gradient(
-    circle at center,
-    rgba(255, 255, 255, 0.1) 0%,
-    transparent 70%
-  );
-  opacity: 0;
-  transition: opacity 0.3s ease;
-}
-
-.home-footer-social-icon:hover::after {
-  opacity: 1;
-}
-
-.home-footer-column {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-4, 16px);
-  min-width: 180px;
-}
-
-.home-footer-col-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0;
-  cursor: default;
-}
-
-.home-footer-col-title {
-  font-size: 14px;
-  font-weight: 700;
-  letter-spacing: 0.05em;
-  margin: 0;
-  color: var(--text-primary, #ffffff);
-}
-
-.home-footer-chevron {
-  display: none;
-}
-
-.home-footer-col-content {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-3, 12px);
-}
-
-.home-footer-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-3, 12px);
-}
-
-.home-footer-link {
-  color: var(--text-muted, rgba(255, 255, 255, 0.6));
-  text-decoration: none;
-  font-size: 14px;
-  transition: color 0.2s ease;
-  display: inline-block;
-  padding: var(--t-space-2, 8px) 0;
-}
-
-.home-footer-link:hover {
-  color: var(--primaryGradientStart, #81f1ff);
-}
-
-.home-footer-bottom-bar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: var(--t-space-6, 24px);
-  margin-top: var(--t-space-12, 48px);
-  padding-top: var(--t-space-8, 32px);
-  border-top: 1px solid var(--borderColor, rgba(255, 255, 255, 0.1));
-}
-
-.home-footer-copyright-col {
-  display: flex;
-  flex-direction: column;
-  gap: var(--t-space-1, 4px);
-}
-
-.home-footer-copyright {
-  font-size: 14px;
-  color: var(--text-muted, rgba(255, 255, 255, 0.6));
-  margin: 0;
-}
-
-.home-footer-version {
-  display: flex;
-  align-items: center;
-  gap: var(--t-space-2, 8px);
-}
-
-.home-footer-status-dot {
-  width: 6px;
-  height: 6px;
-  border-radius: 3px;
-  background-color: var(--color-success, #4ade80);
-}
-
-.home-footer-version-text {
-  font-size: 12px;
-  letter-spacing: 0.05em;
-  color: var(--text-primary, #ffffff);
-}
-
-.home-footer-crafted {
-  font-size: 12px;
-  color: var(--text-primary, #ffffff);
-}
-
-@media (max-width: 768px) {
-  .home-footer-root {
-    padding-top: var(--t-space-10, 40px);
-    padding-bottom: var(--t-space-8, 32px);
-  }
-
-  .home-footer-grid {
-    flex-direction: column;
-    gap: 0;
-    align-items: center;
-  }
-
-  .home-footer-brand-col {
-    flex: none;
-    align-items: center;
-    text-align: center;
-    min-width: 100%;
-    padding-bottom: var(--t-space-10, 40px);
-    border-bottom: 1px solid var(--glass-border, rgba(255, 255, 255, 0.1));
-    margin-bottom: var(--t-space-6, 24px);
-  }
-
-  .home-footer-brand-name,
-  .home-footer-description {
-    text-align: center;
-  }
-
-  .home-footer-social {
-    align-items: center;
-  }
-
-  .home-footer-social-links {
-    justify-content: center;
-  }
-
-  .home-footer-column {
-    min-width: 100%;
-    gap: 0;
-    border-bottom: 1px solid var(--glass-border, rgba(255, 255, 255, 0.1));
-  }
-
-  .home-footer-col-header {
-    padding: var(--t-space-4, 16px) 0;
-    cursor: pointer;
-  }
-
-  .home-footer-chevron {
-    display: block;
-    transition: transform 0.2s ease;
-  }
-
-  .home-footer-column.open .home-footer-chevron {
-    transform: rotate(180deg);
-  }
-
-  .home-footer-col-content {
-    display: none;
-    padding-top: var(--t-space-2, 8px);
-    padding-bottom: var(--t-space-6, 24px);
-    align-items: center;
-  }
-
-  .home-footer-list {
-    align-items: center;
-  }
-
-  .home-footer-column.open .home-footer-col-content {
-    display: flex;
-  }
-
-  .home-footer-bottom-bar {
-    flex-direction: column;
-    align-items: center;
-    margin-top: var(--t-space-10, 40px);
-    text-align: center;
-  }
-
-  .home-footer-copyright-col {
-    align-items: center;
-  }
-}
 /* Static Presentation Styles */
 
 .presentation-container {
