+++ apps/web/src/app/layout.tsx
@@ -7,6 +7,7 @@ import { cookies } from 'next/headers';
 import { appConfig } from '@/shared/config/app-config';
 import { Header } from '@/widgets/header/ui/Header';
 import { AnnouncementBanner } from '@/widgets/AnnouncementBanner/ui/AnnouncementBanner';
+import { LayoutFooter } from '@/widgets/footer';
 import { JsonLd } from '@/shared/ui/JsonLd';
 
 const geistSans = Geist({
@@ -167,6 +168,7 @@ export default async function RootLayout({
                 <AnnouncementBanner />
                 <Header />
                 {children}
+                <LayoutFooter />
                 {authToken ? <WalletLiveBridge authToken={authToken} /> : null}
               </BrowserRegistry>
             </PWAProvider>
