+++ apps/web/src/widgets/footer/index.ts
@@ -1,1 +1,2 @@
 export { default as AppFooter } from './ui/AppFooter';
+export { default as LayoutFooter } from './ui/LayoutFooter';
